requirements1.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
pandas
langchain-groq
validators==0.28.1
unstructured
numexpr
huggingface_hub
streamlit